requirements.txt
asgiref==3.9.1
certifi==2025.8.3
charset-normalizer==3.4.2
Django==5.2.4
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
idna==3.10
mysqlclient==2.2.7
numpy==2.3.2
pandas==2.3.1
pillow==11.3.0
pip==25.2
PyJWT==2.10.1
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.4
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
python-dotenv==1.0.1
pyotp==2.9.0

